+++ src/app/app.component.html
@@ -1,22 +1,2 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Tp Alpine</title>
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body>
-<header><img src="../assets/images/sources-homepage/logo/logo-white.png" alt="logo" class="centered-image"></header>
-  <section>
-
-    <div id="video-container">
-      <video id="video" controls loop>
-        <source src="../assets/video/Alpine-Milan-HomePage-1680x951-950-non-possible-v3-12072021.mp4" type="video/mp4">
-        Your browser does not support the video tag.
-      </video>
-    </div>
-   
-  </section>
-</body>
-</html>
+<app-navbar></app-navbar>
+<app-videocard></app-videocard>
