--- src/app/app.component.scss
@@ -1,30 +1,0 @@
-body {
-  background-color: RGB(48, 48, 48);
-  background-size: cover;
-}
-
-
-header {
- 
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100px; 
-}
-
-.centered-image {
-  display: block;
-  margin: 0;
-}
-
-#video-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 400px;
-}
-
-#video {
-  width: 80%;
-  max-width: 800px;
-}
