--- src/app/color/color.component.html
@@ -1,1 +1,0 @@
-<p>color works!</p>
