+++ src/app/component/color/color.component.html
@@ -1,0 +1,1 @@
+<p>color works!</p>
