+++ src/app/component/footer/footer.component.html
@@ -1,0 +1,1 @@
+<p>footer works!</p>
