+++ src/app/component/navbar/navbar.component.html
@@ -1,0 +1,27 @@
+<nav class="bg-gray-800">
+  <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+    <div class="flex items-center justify-between h-16">
+      <div class="flex items-center">
+        <!-- Logo -->
+        <img src="../../../assets/images/sources-homepage/logo/logo-white.png" alt="logo" class="h-8">
+      </div>
+      <div class="hidden md:block">
+        <ul class="flex space-x-4">
+          <!-- Liens de navigation -->
+          <li><a href="#" class="text-gray-300 hover:text-white">Accueil</a></li>
+          <li><a href="#" class="text-gray-300 hover:text-white">À propos</a></li>
+          <li><a href="#" class="text-gray-300 hover:text-white">Services</a></li>
+          <li><a href="#" class="text-gray-300 hover:text-white">Contact</a></li>
+        </ul>
+      </div>
+    </div>
+  </div>
+</nav>
+
+<header>
+  <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-4">
+    <img src="../../../assets/images/sources-homepage/logo/logo-white.png" alt="logo" class="h-8">
+  </div>
+</header>
+
+
