+++ src/app/component/rim/rim.component.html
@@ -1,0 +1,1 @@
+<p>rim works!</p>
