+++ src/app/component/rim/rim.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-rim',
+  templateUrl: './rim.component.html',
+  styleUrls: ['./rim.component.scss']
+})
+export class RimComponent {
+
+}
