+++ src/app/component/version/version.component.html
@@ -1,0 +1,1 @@
+<p>version works!</p>
