+++ src/app/component/version/version.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-version',
+  templateUrl: './version.component.html',
+  styleUrls: ['./version.component.scss']
+})
+export class VersionComponent {
+
+}
