+++ src/app/component/videocard/summary.component.html
@@ -1,0 +1,17 @@
+<section class="bg-gray-100 py-12">
+  <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 flex items-center">
+    <div class="w-full md:w-1/2">
+      <div id="video-container" class="relative">
+        <video id="video" autoplay loop class="w-full">
+          <source src="../../../assets/video/Alpine-Milan-HomePage-1680x951-950-non-possible-v3-12072021.mp4" type="video/mp4">
+          Votre navigateur ne prend pas en charge la balise vidéo.
+        </video>
+      </div>
+    </div>
+    <div class="w-full md:w-1/2 px-6">
+      <h2 class="text-3xl font-bold mb-4">Alpine : Le plaisir de conduire</h2>
+      <p class="text-gray-700">Découvrez la marque Alpine, une icône du sport automobile français. Depuis des décennies, Alpine crée des voitures légères, agiles et performantes, offrant une expérience de conduite inégalée. Avec leur design distinctif et leur héritage riche, les véhicules Alpine incarnent le plaisir de conduire.</p>
+    </div>
+  </div>
+</section>
+
