+++ src/app/component/videocard/summary.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-videocard',
+  templateUrl: './summary.component.html',
+  styleUrls: ['./summary.component.scss']
+})
+export class SummaryComponent {
+
+}
