--- src/app/rim/rim.component.html
@@ -1,1 +1,0 @@
-<p>rim works!</p>
