--- src/app/summary/summary.component.html
@@ -1,1 +1,0 @@
-<p>summary works!</p>
