--- src/app/version/version.component.html
@@ -1,1 +1,0 @@
-<p>version works!</p>
