+++ src/styles.scss
@@ -1,1 +1,6 @@
 /* You can add global styles to this file, and also import other style files */
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+body { margin : 0;}
