+++ tailwind.config.js
@@ -1,0 +1,10 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    "./src/**/*.{html,ts, scss}",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
